requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
Faker==29.0.0
filelock==3.15.4
identify==2.6.0
Markdown==3.7
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.8.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
virtualenv==20.26.3